+++ node/src/template_chain_spec.rs
@@ -1,9 +1,9 @@
 use crate::chain_spec::*;
 use chain_params::SidechainParams;
 use sc_service::ChainType;
-use sidechain_domain::{MainchainAddressHash, UtxoId};
+use sidechain_domain::{AssetName, MainchainAddress, MainchainAddressHash, PolicyId, UtxoId};
 use sidechain_runtime::{
-	AuraConfig, BalancesConfig, GrandpaConfig, RuntimeGenesisConfig,
+	AuraConfig, BalancesConfig, GrandpaConfig, NativeTokenManagementConfig, RuntimeGenesisConfig,
 	SessionCommitteeManagementConfig, SessionConfig, SidechainConfig, SudoConfig, SystemConfig,
 };
 
@@ -43,6 +43,18 @@ pub fn chain_spec() -> Result<ChainSpec, EnvVarReadError> {
 			// Same as SessionConfig
 			initial_authorities: vec![],
 			main_chain_scripts: read_mainchain_scripts_from_env()?,
+		},
+		native_token_management: NativeTokenManagementConfig {
+			main_chain_scripts: sp_native_token_management::MainChainScripts {
+				native_token_policy: from_var::<PolicyId>("PARTNER_CHAIN_NATIVE_TOKEN_POLICY_ID")?,
+				native_token_asset_name: from_var::<AssetName>(
+					"PARTNER_CHAIN_NATIVE_TOKEN_ASSET_NAME",
+				)?,
+				illiquid_supply_address: from_var::<MainchainAddress>(
+					"PARTNER_CHAIN_ILLIQUID_SUPPLY_VALIDATOR_ADDRESS",
+				)?,
+			},
+			..Default::default()
 		},
 	};
 	let genesis_json = serde_json::to_value(runtime_genesis_config)